Add comprehensive functional tests for conversation streams
